`@remotion/renderer`: Split up into multiple functions
Will make the next refactor easier
