+++ packages/renderer/src/cycle-browser-tabs.ts
@@ -2,12 +2,17 @@ import type {LogLevel} from './log-level';
 import {Log} from './logger';
 import type {BrowserReplacer} from './replace-browser';
 
-export const cycleBrowserTabs = (
-	puppeteerInstance: BrowserReplacer,
-	concurrency: number,
-	logLevel: LogLevel,
-	indent: boolean,
-): {
+export const cycleBrowserTabs = ({
+	puppeteerInstance,
+	concurrency,
+	logLevel,
+	indent,
+}: {
+	puppeteerInstance: BrowserReplacer;
+	concurrency: number;
+	logLevel: LogLevel;
+	indent: boolean;
+}): {
 	stopCycling: () => void;
 } => {
 	if (concurrency <= 1) {
